Added Character Stat Modification GUI for Admins (#2)
* Moved some things around

Just moved some things for order.

* Added Character Stat Modification GUI

Very long commit description here:

The GUI contains four buttons, as well as a choice of five
radio-buttons.
-The first button is Pickup. When activated, the admit is able to pickup
every item on screen at max zoom by pressing F. (WARNING - This includes
items on transport belts, so be very careful with this)
-The second button is Reach. This extends the player's reach for all
aspects: building, item dropping, structure interaction, and ore/tree
mining. Same as above, will be the entire screen at max zoom level
-The third button is Crafting. This will increase the crafting speed to
be 60 times faster, meaning most items are crafted instantly.
-The fourth button is Mining. This makes mining ores, chopping trees,
and deconstructing buildings. The speed increase makes it near instant.
The five radio buttons are only active one at a time. They control the
run speed. 1x is default run speed, 2x is twice as fast. There's also
3x, 5x, and 10x.

The update also activates cheat_mode for admins who enter the god
character controller mode (Spectate mode). This allows crafting to be
instant and not cost resources.
Character settings are saved when the admin enters Spectate mode, and
will be restored to the new character when exiting Spectate mode.

* Fixed Bugs

Missed quite a few bugs because I didn't test before committing...
Whoops.

* Disabled Button if in Spectate Mode

Meant to do this earlier, never actually put the code in.
Bugs occur if the admin is in Spectate mode while attempting to modify
character. Disabled the button so that it couldn't be activated while in
Spectate mode.

* Fixed Typo

This is the dangers of copy/pasting

* Final Minor Fixes

Spacing, an extra Label element on the GUI