Follow Specific Player (#3)
* Add initial follow functionality

* Trying to add a return option

* Fix bugs

* Remove players own name from follow list

* Fix disconnect bug

If player disconnects while he is being followed, it might result in a crash.
This commit should fix it

* Few changes

* Corrected Bugs

Return did not work as intended, potential bug when connected players
would change

* Standardized whitespace

There were mixed occurances of TAB and 4 spaces throughout this program,
and it was messing with GitHub's viewer (See Discord code chat for
example). I changed all occurances of 4 spaces to TAB to correct this.